Use unique identifiers for sockets
This commit modifies the plugin binary to accept a single argument as a socket
name which is generated upon calling `plugin.Use`. This now allows multiple instances
of a plugin to be used on one machine (if that's desirable) and avoids possible clashes
where plugins may not have gracefully exited.

Signed-off-by: David Bond <davidsbond93@gmail.com>